+++ arviz/tests/external_tests/test_data_pymc.py
@@ -10,6 +10,7 @@ from numpy import ma
 
 from arviz import (  # pylint: disable=wrong-import-position
     InferenceData,
+    from_dict,
     from_pymc3,
     from_pymc3_predictions,
 )
@@ -107,7 +108,7 @@ class TestDataPyMC3:
                 )
 
     def test_from_pymc_predictions(self, data, eight_schools_params):
-        "Test that we can add predictions to a previously-existing InferenceData."
+        """Test that we can add predictions to a previously-existing InferenceData."""
         test_dict = {
             "posterior": ["mu", "tau", "eta", "theta"],
             "sample_stats": ["diverging", "lp"],
@@ -139,6 +140,15 @@ class TestDataPyMC3:
             assert ivalues.shape[0] == 1  # one chain in predictions
             assert np.all(np.isclose(ivalues[0], values))
 
+    def test_from_pymc_trace_inference_data(self):
+        """Check if the error is raised successfully after passing InferenceData as trace"""
+        idata = from_dict(
+            posterior={"A": np.random.randn(2, 10, 2), "B": np.random.randn(2, 10, 5, 2)}
+        )
+        assert isinstance(idata, InferenceData)
+        with pytest.raises(ValueError):
+            from_pymc3(trace=idata, model=pm.Model())
+
     def test_from_pymc_predictions_new(self, data, eight_schools_params):
         # check creating new
         inference_data, posterior_predictive = self.make_predictions_inference_data(
@@ -147,7 +157,7 @@ class TestDataPyMC3:
         test_dict = {
             "posterior": ["mu", "tau", "eta", "theta"],
             "predictions": ["obs"],
-            "~observed_data": "",
+            "~observed_data": [""],
         }
         fails = check_multiple_attrs(test_dict, inference_data)
         assert not fails
@@ -329,7 +339,7 @@ class TestDataPyMC3:
         }
         if not log_likelihood:
             test_dict.pop("log_likelihood")
-            test_dict["~log_likelihood"] = []
+            test_dict["~log_likelihood"] = [""]
         if isinstance(log_likelihood, list):
             test_dict["log_likelihood"] = ["y1", "~y2"]
 
@@ -453,7 +463,7 @@ class TestDataPyMC3:
             # assert predictive_trace["obs"].shape == (400, 2)
             # but the shape seems to vary between pymc3 versions
             inference_data = from_pymc3_predictions(predictive_trace, posterior_trace=trace)
-        test_dict = {"posterior": ["beta"], "~observed_data": ""}
+        test_dict = {"posterior": ["beta"], "~observed_data": [""]}
         fails = check_multiple_attrs(test_dict, inference_data)
         assert not fails, "Posterior data not copied over as expected."
         test_dict = {"predictions": ["obs"]}
@@ -530,7 +540,7 @@ class TestDataPyMC3:
             inference_data = from_pymc3(prior=prior)
         test_dict = {
             "prior": ["beta", "obs"],
-            "~prior_predictive": [],
+            "~prior_predictive": [""],
         }
         fails = check_multiple_attrs(test_dict, inference_data)
         assert not fails
@@ -583,8 +593,8 @@ class TestPyMC3WarmupHandling:
             f"{post_prefix}sample_stats": ["~tune", "accept"],
             f"{warmup_prefix}warmup_posterior": ["u1", "n1"],
             f"{warmup_prefix}warmup_sample_stats": ["~tune"],
-            "~warmup_log_likelihood": [],
-            "~log_likelihood": [],
+            "~warmup_log_likelihood": [""],
+            "~log_likelihood": [""],
         }
         fails = check_multiple_attrs(test_dict, idata)
         assert not fails
@@ -621,8 +631,8 @@ class TestPyMC3WarmupHandling:
         test_dict = {
             "posterior": ["u1", "n1"],
             "sample_stats": ["~tune", "accept"],
-            "~warmup_posterior": [],
-            "~warmup_sample_stats": [],
+            "~warmup_posterior": [""],
+            "~warmup_sample_stats": [""],
         }
         fails = check_multiple_attrs(test_dict, idata)
         assert not fails
@@ -667,8 +677,8 @@ class TestPyMC3WarmupHandling:
             test_dict = {
                 "posterior": ["u1", "n1"],
                 "sample_stats": ["~tune", "accept"],
-                "~warmup_posterior": [],
-                "~warmup_sample_stats": [],
+                "~warmup_posterior": [""],
+                "~warmup_sample_stats": [""],
             }
             fails = check_multiple_attrs(test_dict, idata)
             assert not fails